box {bbb}	R Documentation

Build a bounding box

Description

Launches an interactive shiny session for you to draw a bounding box. The code to recreate the bounding box is then available in your script.

Usage

box()

[Package bbb version 1.0.0 Index]