manual_bbox {bbb}	R Documentation

Manually build a bounding box using the Shiny rectangle

Description

Manually construct the bounding box and return a vector.

Usage

manual_bbox(coords)

Arguments

coords
from the shiny

Value

a matrix

[Package bbb version 1.0.0 Index]