morph {bbb}	R Documentation

Morph the format of a bounding box

Description

Change the layout and names of a bounding box. The function will figure out the current layout of your bounding box and morph it to the desired output style.

Usage

morph(bbox, output_type)

Arguments

bbox
a bounding box

output_type
(quoted character) the desired output format. One of "xy_string", "word_string" or "xy_matrix".

Value

a bounding box in the desired format

Examples

## Not run: 
bbox <- c("left" = -73.848849, "bottom" = -40.979898, "right" = -28.11848, "top" = 0.703107)
attr(bbox, "class") <- "bbox"
bbox2 <- morph(bbox, "xy_string")

## End(Not run)

[Package bbb version 1.0.0 Index]