word_str_to_xy_matrix {bbb}	R Documentation

Morph an xy string bbox to a xy matrix bbox

Description

Changes the format of a xy string bounding box (eg xmax, ymin) to an xy matrix type.

Usage

word_str_to_xy_matrix(bbox)

Arguments

bbox
A bounding box with names xmin, ymax etc

Value

a bounding box in matrix form

[Package bbb version 1.0.0 Index]